[#78] dto 계층 파일 분리
